requirements.txt
Django>=2.0
django-bootstrap4
django-registration
django-templated-email
django-crontab
